build: split HttpSession (dashboard/REST router) into its own TU (#735 slice 1)
First mechanical slice of the web_server source-dissolution (#735, follow-up
to the #63 object-lib refactor). Moves the HTTP request router
core::HttpSession — the dashboard / REST / SSE presentation surface — and its
four file-local URL/validation helpers (url_decode, is_valid_hex_hash,
is_valid_address_chars, is_valid_graph_param) out of the 9.1k-line
web_server.cpp into a new translation unit src/core/http_session.cpp.

Rationale: web_server.cpp conflates two concerns that #735 wants separated —
the miner-facing engine (core::MiningInterface + the StratumServer it owns;
stratum_server.cpp depends on MiningInterface) and the operator-facing
dashboard. HttpSession is a pure consumer of MiningInterface's public
rest_*() accessors; it holds no mining state. Isolating it in its own TU
narrows that seam and is the smallest safe step toward the eventual
core/web target split.

Behaviour-neutral: the moved code is byte-identical to its former location
(verified: the 864 removed lines equal the new TU's body). to_hex() stays in
web_server.cpp (used only by MiningInterface). Symbol audit confirms all six
HttpSession methods now resolve from http_session.cpp.o and none remain in
web_server.cpp.o. The core object library compiles and links clean
(Release, Conan 2).

No consensus surface touched.
